// src/ApiModule/model/scheduleRecord.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Program } from './program';


export interface ScheduleRecord { 
    program: Program;
    date: string;
    duration: number;
    registration_opens_at?: string;
    places_available?: number;
}