// src/ApiModule/model/schemaRecord.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Program } from './program';


export interface SchemaRecord { 
    week_day: number;
    day_time: string;
    duration: number;
    id: number;
    program: Program;
}